# Presentation: Hierarchy filtering perf. improvements