… iModelCore/ECPresentation/Source/PresentationManagerImpl.h
… iModelCore/ECPresentation/Tests/NonPublished/Helpers/TestRulesDrivenECPresentationManagerImpl.h
… iModelCore/ECPresentation/Tests/NonPublished/Unit/Hierarchies/NodesProviderTests.cpp