Filtered models tree filtering performance on a specific test iModel (17 GB, lots of Subjects & Models, 1281 nodes in filtered hierarchy):

| | 4.0.12 | before this PR | after this PR |
| --- | --- | --- | --- |
| Initial filter | 59.453 s | 47.179 s | 44.385 s |
| Subsequent filters | 3.671 s | 2.277 s | 1.429 s |

Not sure if we can squeeze anything more from this...

itwinjs-core PR: https://github.com/iTwin/itwinjs-core/pull/5767
itwinjs-core: https://github.com/iTwin/itwinjs-core/tree/presentation/update-snapshots-after-filtering-changes